refactor: change structure of intro-page files/change events structure/replace html basic layout introPageRender function
