--- src/js/components/nav-menu.js
@@ -1,5 +1,0 @@
-
-
-export default function menuButtonHandler () {
-  this.classList.toggle('is-active');
-};
